Create prototype for applying tracking to memoizer
A class can be provided to TrackingMemoizer.get_tracking_instance, which
creates a new class and instance thereof with tracking capabilities.

This method is somewhat hacky - better options are...
A) Add tracking to the base Memoizer instead
- This increases the memory cost for functions with simple args and
returns not-insubstantially, but is that actually going to be relevant?
- If this route is taken, make sure to store params/results in one dict
B) Create a base Template class and a TemplateFactory utility class
- TemplateFactory can provide methods for applying Templates to classes
- This *should* be feasible without metaclasses, but investigate
- More flexible, more complex, more elegant, more hacky
- Consider: What other Templates would be created? Is it worth doing?